third_party/blink/web_tests/external/wpt/css/css-scroll-anchoring/anchor-inside-textarea.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Scroll anchoring works on textarea
  assert_equals: The scroll position should be adjusted expected 310 but got 300
Harness: the test ran to completion.
